move stop/start into nodes and simplify endpoint move API
