+++ nodes/default_node.go
@@ -11,9 +11,11 @@ import (
 	"path"
 	"path/filepath"
 	"regexp"
+	goruntime "runtime"
 	"strconv"
 	"strings"
 	"sync"
+	"time"
 
 	"github.com/charmbracelet/log"
 	"github.com/containernetworking/plugins/pkg/ns"
@@ -26,6 +28,7 @@ import (
 	clabtypes "github.com/srl-labs/containerlab/types"
 	clabutils "github.com/srl-labs/containerlab/utils"
 	"github.com/vishvananda/netlink"
+	"github.com/vishvananda/netns"
 )
 
 const (
@@ -819,4 +822,399 @@ func (d *DefaultNode) GetContainerStatus(ctx context.Context) clabruntime.Contai
 
 func (d *DefaultNode) IsHealthy(ctx context.Context) (bool, error) {
 	return d.Runtime.IsHealthy(ctx, d.GetContainerName())
+}
+
+// ---------------------------------------------------------------------------
+// Lifecycle: Stop / Start
+// ---------------------------------------------------------------------------
+
+const vrnetlabVersionLabel = "vrnetlab-version"
+
+var imageTagRE = regexp.MustCompile(`^[A-Za-z0-9][A-Za-z0-9._-]*$`)
+
+// Stop parks dataplane interfaces into a dedicated network namespace and stops the container.
+func (d *DefaultNode) Stop(ctx context.Context) error {
+	cfg := d.Config()
+
+	parkName := clabutils.ParkingNetnsName(cfg.LongName)
+	parkPath, err := ensureNamedNetns(parkName)
+	if err != nil {
+		return fmt.Errorf("node %q failed creating parking netns: %w", cfg.ShortName, err)
+	}
+	parkingNode := clablinks.NewGenericLinkNode(parkName, parkPath)
+
+	nodeNSPath, err := d.GetNSPath(ctx)
+	if err != nil {
+		return fmt.Errorf("node %q failed getting netns path: %w", cfg.ShortName, err)
+	}
+
+	// Move dataplane interfaces into the parking netns while the container netns is still alive.
+	moved, err := moveEndpoints(
+		d.GetEndpoints(),
+		func(ep clablinks.Endpoint) error {
+			return ep.MoveTo(ctx, parkingNode, preMoveSetDownOptions())
+		},
+	)
+	if err != nil {
+		// Roll back any endpoints already moved to the parking namespace.
+		if len(moved) > 0 {
+			if rbErr := rollbackEndpoints(moved, func(ep clablinks.Endpoint) error {
+				return moveEndpointToOwner(ctx, ep, parkingNode, nil)
+			}); rbErr != nil {
+				return fmt.Errorf(
+					"node %q failed parking interfaces: %w (rollback failed: %v)",
+					cfg.ShortName,
+					err,
+					rbErr,
+				)
+			}
+			_ = setEndpointsUp(ctx, moved)
+		}
+		return fmt.Errorf("node %q failed parking interfaces: %w", cfg.ShortName, err)
+	}
+
+	// Repoint /run/netns/<containerName> to the parking netns so that inspect/destroy keep working.
+	if err := clabutils.LinkContainerNS(parkPath, cfg.LongName); err != nil {
+		return fmt.Errorf("node %q failed linking parking netns: %w", cfg.ShortName, err)
+	}
+
+	preStopCleanup(ctx, d)
+
+	if err := d.Runtime.StopContainer(ctx, cfg.LongName); err != nil {
+		// Docker/podman may return an error while the container is already stopped (timeout, API hiccup).
+		// Treat this as success if the desired state is reached.
+		status := d.Runtime.GetContainerStatus(ctx, cfg.LongName)
+		if status == clabruntime.Stopped {
+			log.Warnf("node %q stop returned error but container is stopped: %v", cfg.ShortName, err)
+			return nil
+		}
+
+		// Roll back to a running node with interfaces restored.
+		if linkErr := clabutils.LinkContainerNS(nodeNSPath, cfg.LongName); linkErr != nil {
+			log.Warnf("node %q failed restoring /run/netns symlink after stop error: %v", cfg.ShortName, linkErr)
+		}
+		if rbErr := rollbackEndpoints(moved, func(ep clablinks.Endpoint) error {
+			return moveEndpointToOwner(ctx, ep, parkingNode, nil)
+		}); rbErr != nil {
+			return fmt.Errorf(
+				"node %q failed stopping container: %w (rollback failed restoring interfaces: %v)",
+				cfg.ShortName,
+				err,
+				rbErr,
+			)
+		}
+		_ = setEndpointsUp(ctx, moved)
+
+		return fmt.Errorf("node %q failed stopping container: %w", cfg.ShortName, err)
+	}
+
+	return nil
+}
+
+// Start restarts a stopped container and restores its parked dataplane interfaces.
+func (d *DefaultNode) Start(ctx context.Context) error {
+	cfg := d.Config()
+
+	parkName := clabutils.ParkingNetnsName(cfg.LongName)
+	parkPath := filepath.Join("/run/netns", parkName)
+	if !clabutils.FileOrDirExists(parkPath) {
+		return fmt.Errorf(
+			"node %q has no parking netns %q; seamless start requires stopping via containerlab",
+			cfg.ShortName,
+			parkName,
+		)
+	}
+	parkingNode := clablinks.NewGenericLinkNode(parkName, parkPath)
+
+	if _, err := d.Runtime.StartContainer(ctx, cfg.LongName, d); err != nil {
+		return fmt.Errorf("node %q failed starting container: %w", cfg.ShortName, err)
+	}
+
+	if _, err := d.GetNSPath(ctx); err != nil {
+		// Try to keep destroy/inspect operational by repointing back to the parking netns.
+		_ = clabutils.LinkContainerNS(parkPath, cfg.LongName)
+		_ = d.Runtime.StopContainer(ctx, cfg.LongName)
+		return fmt.Errorf("node %q failed getting netns path: %w", cfg.ShortName, err)
+	}
+
+	// Move interfaces back into the container netns.
+	moved, err := moveEndpoints(d.GetEndpoints(), func(ep clablinks.Endpoint) error {
+		return moveEndpointToOwner(ctx, ep, parkingNode, nil)
+	})
+	if err != nil {
+		// Attempt rollback to keep the node in a consistent stopped+parked state.
+		_ = rollbackEndpoints(moved, func(ep clablinks.Endpoint) error {
+			return ep.MoveTo(ctx, parkingNode, preMoveSetDownOptions())
+		})
+		_ = clabutils.LinkContainerNS(parkPath, cfg.LongName)
+		_ = d.Runtime.StopContainer(ctx, cfg.LongName)
+		return fmt.Errorf("node %q failed restoring interfaces: %w", cfg.ShortName, err)
+	}
+
+	// Bring restored interfaces up.
+	if err := setEndpointsUp(ctx, moved); err != nil {
+		// Attempt rollback to keep the node in a consistent stopped+parked state.
+		_ = rollbackEndpoints(moved, func(ep clablinks.Endpoint) error {
+			return ep.MoveTo(ctx, parkingNode, preMoveSetDownOptions())
+		})
+		_ = clabutils.LinkContainerNS(parkPath, cfg.LongName)
+		_ = d.Runtime.StopContainer(ctx, cfg.LongName)
+		return fmt.Errorf("node %q failed enabling interfaces: %w", cfg.ShortName, err)
+	}
+
+	// Re-run topology exec commands on lifecycle start to restore node-local interface config
+	// (for example IP addresses added during deploy exec phase).
+	execCollection := clabexec.NewExecCollection()
+	if err := d.RunExecFromConfig(ctx, execCollection); err != nil {
+		log.Errorf("failed to run exec commands for node %q on lifecycle start: %v", cfg.ShortName, err)
+	}
+	execCollection.Log()
+
+	return nil
+}
+
+// ---------------------------------------------------------------------------
+// Lifecycle helpers (unexported)
+// ---------------------------------------------------------------------------
+
+func ensureNamedNetns(name string) (string, error) {
+	nspath := filepath.Join("/run/netns", name)
+	if clabutils.FileOrDirExists(nspath) {
+		return nspath, nil
+	}
+
+	goruntime.LockOSThread()
+	defer goruntime.UnlockOSThread()
+
+	origNS, err := netns.Get()
+	if err != nil {
+		return "", err
+	}
+	defer origNS.Close()
+	defer func() {
+		_ = netns.Set(origNS)
+	}()
+
+	newNS, err := netns.NewNamed(name)
+	if err != nil {
+		if os.IsExist(err) && clabutils.FileOrDirExists(nspath) {
+			return nspath, nil
+		}
+		return "", err
+	}
+	newNS.Close()
+
+	return nspath, nil
+}
+
+func moveEndpointToOwner(
+	ctx context.Context,
+	ep clablinks.Endpoint,
+	src clablinks.Node,
+	opts *clablinks.MoveOptions,
+) error {
+	srcEp := clablinks.NewEndpointGeneric(src, ep.GetIfaceName(), ep.GetLink())
+	srcEp.SetIfaceAlias(ep.GetIfaceAlias())
+
+	return srcEp.MoveTo(ctx, ep.GetNode(), opts)
+}
+
+func preMoveSetDownOptions() *clablinks.MoveOptions {
+	return &clablinks.MoveOptions{
+		PreMove: netlink.LinkSetDown,
+	}
+}
+
+func moveEndpoints(
+	endpoints []clablinks.Endpoint,
+	move func(clablinks.Endpoint) error,
+) ([]clablinks.Endpoint, error) {
+	moved := make([]clablinks.Endpoint, 0, len(endpoints))
+
+	for _, ep := range endpoints {
+		if err := move(ep); err != nil {
+			return moved, err
+		}
+		moved = append(moved, ep)
+	}
+
+	return moved, nil
+}
+
+func rollbackEndpoints(
+	moved []clablinks.Endpoint,
+	move func(clablinks.Endpoint) error,
+) error {
+	for i := len(moved) - 1; i >= 0; i-- {
+		if err := move(moved[i]); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func setEndpointsUp(ctx context.Context, endpoints []clablinks.Endpoint) error {
+	for _, ep := range endpoints {
+		if err := ep.SetUp(ctx); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func preStopCleanup(ctx context.Context, d *DefaultNode) {
+	if isVrnetlabNode(ctx, d) {
+		preStopPrepareVrnetlabQcowAlias(ctx, d)
+	}
+
+	preStopCleanupNamedNetns(ctx, d)
+}
+
+func preStopCleanupNamedNetns(ctx context.Context, d *DefaultNode) {
+	// Best-effort cleanup for containers that create named network namespaces under /run/netns.
+	// We lazily unmount active nsfs mounts and remove stale entries to avoid namespace artifacts
+	// breaking subsequent starts.
+	ctx, cancel := context.WithTimeout(ctx, 10*time.Second)
+	defer cancel()
+
+	cmd := `if [ -d /run/netns ]; then ` +
+		`awk '$5 ~ "^/run/netns/" {print $5}' /proc/self/mountinfo 2>/dev/null | ` +
+		`while IFS= read -r mp; do ` +
+		`umount -l "$mp" 2>/dev/null || true; ` +
+		`rm -f "$mp" 2>/dev/null || true; ` +
+		`done; ` +
+		`for f in /run/netns/*; do [ -e "$f" ] || break; rm -f "$f" 2>/dev/null || true; done; ` +
+		`fi`
+
+	execCmd := clabexec.NewExecCmdFromSlice([]string{"sh", "-lc", cmd})
+	if res, err := d.RunExec(ctx, execCmd); err != nil {
+		log.Debugf(
+			"node %q generic pre-stop named-netns cleanup skipped/failed: %v",
+			d.Config().ShortName,
+			err,
+		)
+	} else if res != nil && res.ReturnCode != 0 {
+		log.Debugf(
+			"node %q generic pre-stop named-netns cleanup returned code %d (stderr: %s)",
+			d.Config().ShortName,
+			res.ReturnCode,
+			res.Stderr,
+		)
+	}
+}
+
+func isVrnetlabNode(ctx context.Context, d *DefaultNode) bool {
+	containers, err := d.GetContainers(ctx)
+	if err != nil {
+		log.Debugf(
+			"node %q vrnetlab detection skipped: failed to get container metadata: %v",
+			d.Config().ShortName,
+			err,
+		)
+		return false
+	}
+
+	for _, container := range containers {
+		if _, ok := container.Labels[vrnetlabVersionLabel]; ok {
+			return true
+		}
+	}
+
+	return false
+}
+
+func preStopPrepareVrnetlabQcowAlias(ctx context.Context, d *DefaultNode) {
+	aliasName, ok := vrnetlabQcowAliasName(d.Config().Image)
+	if !ok {
+		log.Debugf(
+			"node %q pre-stop vrnetlab qcow alias skipped: unable to infer tag from image %q",
+			d.Config().ShortName,
+			d.Config().Image,
+		)
+		return
+	}
+
+	ctx, cancel := context.WithTimeout(ctx, 10*time.Second)
+	defer cancel()
+
+	// Some vrnetlab nodes rename the original versioned qcow image after first boot and fail on
+	// subsequent starts when they try to rediscover a versioned qcow filename. If there is exactly
+	// one non-overlay qcow file in / and our alias is absent, create a hardlink alias based on the
+	// image tag.
+	cmd := fmt.Sprintf(
+		`alias="/%s"; `+
+			`[ -e "$alias" ] && exit 0; `+
+			`src=""; `+
+			`if [ -f /sros.qcow2 ] && [ "/sros.qcow2" != "$alias" ]; then `+
+			`src="/sros.qcow2"; `+
+			`else `+
+			`set -- /*.qcow2; `+
+			`if [ "$1" != "/*.qcow2" ]; then `+
+			`for f in "$@"; do `+
+			`[ "$f" = "$alias" ] && continue; `+
+			`base="${f##*/}"; `+
+			`case "$base" in *overlay*.qcow2) continue ;; esac; `+
+			`if [ -n "$src" ]; then src=""; break; fi; `+
+			`src="$f"; `+
+			`done; `+
+			`fi; `+
+			`fi; `+
+			`[ -n "$src" ] || exit 0; `+
+			`ln "$src" "$alias"`,
+		aliasName,
+	)
+
+	execCmd := clabexec.NewExecCmdFromSlice([]string{"sh", "-lc", cmd})
+	res, err := d.RunExec(ctx, execCmd)
+	if err != nil {
+		log.Warnf(
+			"node %q pre-stop vrnetlab qcow alias preparation failed: %v",
+			d.Config().ShortName,
+			err,
+		)
+		return
+	}
+
+	if res != nil && res.ReturnCode != 0 {
+		log.Warnf(
+			"node %q pre-stop vrnetlab qcow alias prep returned code %d (stderr: %s)",
+			d.Config().ShortName,
+			res.ReturnCode,
+			res.Stderr,
+		)
+	}
+}
+
+func vrnetlabQcowAliasName(image string) (string, bool) {
+	tag, ok := imageTag(image)
+	if !ok {
+		return "", false
+	}
+
+	return "clab-" + tag + ".qcow2", true
+}
+
+func imageTag(image string) (string, bool) {
+	if at := strings.LastIndex(image, "@"); at != -1 {
+		image = image[:at]
+	}
+
+	lastSlash := strings.LastIndex(image, "/")
+	lastColon := strings.LastIndex(image, ":")
+	if lastColon == -1 || lastColon < lastSlash {
+		return "", false
+	}
+
+	tag := image[lastColon+1:]
+	if tag == "" {
+		return "", false
+	}
+
+	if !imageTagRE.MatchString(tag) {
+		return "", false
+	}
+
+	return tag, true
 }
